+++ projects/astrolend/index.js
@@ -1,34 +1,32 @@
 const { Program } = require("@coral-xyz/anchor");
-const { getProvider, sumTokens2, } = require("../helper/solana");
+const { getProvider, sumTokens2 } = require("../helper/solana");
+const idl = require('./idl.json')
+const wrappedI80F48toBigNumber = require("./utils/conversion")
 
-const idl = require('./idl');
+let _banks
 
-const utils = require("./utils/conversion");
-const GROUP_MAINNET= "8GzZHDKts3oHeL91h4fYjbjaAcUicBb8NB6ZTLTHvYr6"
-async function tvl() {
-    const provider = getProvider("eclipse")
-    const program = new Program(idl, provider)
-   
-    const banks= (await program.account.bank.all()).filter((bank)=>{
-      return bank.account.group.toString()== GROUP_MAINNET
-     })
-    
-        const map = banks.reduce((acc, bank) => {
-        // Convert the mint to a string to use it as an object key.
-        const mintString = 'eclipse:'+bank.account.mint.toString();
-      
-        // Set the object property to the numeric value you want.
-        acc[mintString] = utils.conversion.wrappedI80F48toBigNumber(
-          bank.account.totalAssetShares.value
-        );
-      
-        return acc;
-      }, {});
-   return map
-   
+async function getBanks() {
+  if (_banks) return _banks
+  const provider = getProvider('eclipse')
+  const program = new Program(idl, provider)
+  _banks = program.account.bank.all()
+  return _banks
+}
+
+async function tvl(api) {
+  const banks = await getBanks()
+  return sumTokens2({ api, tokenAccounts: banks.map(bank => bank.account.liquidityVault) });
+}
+
+async function borrowed(api) {
+  const banks = await getBanks()
+
+  banks.forEach(bank => {
+    api.add(bank.account.mint.toString(), wrappedI80F48toBigNumber(bank.account.totalLiabilityShares.value))
+  })
 }
 
 module.exports = {
   timetravel: false,
-  solana: { tvl, },
+  eclipse: { tvl, borrowed, },
 }
